CI: disable docker image build and push from a fork
CI running on forks uses docker images from the Scilab Docker registry instead of the fork one.

registry.gitlab.com/**scilab**/scilab instead of registry.gitlab.com/**contributor_name**/scilab

In this way, the images used are the same for everyone.

But, when a contributor pushes changes that will update docker images this will make 2 images on the Scilab Docker registry.
The one created from the fork pipeline and the one made by the MR pipeline.
This MR disables the creation of the Docker image created by the fork that does not aim to be used in the CI.

See https://gitlab.com/scilab/scilab/-/merge_requests/965