fix #17279 - double delete fixed
The **result **was not reset to null, so the **result **was deleted at the RunVisitor deletion and the scope end.
This is why there were no crashes when the structure was declared outside the function test (no scope end) but Scilab probably crashed at exit.

Closes #17279
See https://gitlab.com/scilab/scilab/-/merge_requests/1007